Fix security vulnerabilities and upgrade Tomcat to 11.0.22

+++ web/src/main/java/com/law/admin/controller/ConsultationController.java
@@ -73,7 +73,8 @@ public class ConsultationController {
         Integer total = jdbc.queryForObject(
                 "SELECT COUNT(*) FROM lw_consultation" + where, Integer.class, params.toArray());
 
-        int offset = Math.max((page - 1) * size, 0);
+        // CWE-682：使用 long 運算避免大頁碼導致整數溢位
+        long offset = Math.max((long)(page - 1) * size, 0L);
         List<Object> queryParams = new ArrayList<>(params);
         queryParams.add(size);
         queryParams.add(offset);
